requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
Babel==2.14.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
bokeh==3.3.4
Brlapi==0.8.3
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.1
command-not-found==0.3
contourpy==1.2.0
cryptography==41.0.7
cupshelpers==1.0
dataclasses-json==0.6.3
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.8.0
distro-info===1.1build1
Django==5.0.1
djangorestframework==3.14.0
dnspython==2.4.2
duplicity==0.8.21
email-validator==2.1.0.post1
exception==0.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.105.0
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.6.0
fqdn==1.5.1
frozenlist==1.4.0
future==0.18.2
greenlet==3.0.1
grpcio==1.60.1
grpcio-health-checking==1.60.1
grpcio-tools==1.60.1
gyp==0.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
importlib-metadata==4.6.4
ipykernel==6.29.2
ipython==8.21.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==23.5.0
langchain==0.0.345
langchain-core==0.0.9
langsmith==0.0.69
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
linkify-it-py==2.0.3
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
netifaces==0.11.0
notebook_shim==0.2.3
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
openai==0.28.0
orjson==3.9.10
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
panel==1.3.8
param==2.0.2
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pyclean==2.7.6
pycparser==2.21
pycups==2.0.1
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
Pygments==2.17.2
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2022.1
pyviz_comms==3.0.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.10.3
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
systemd-python==234
tenacity==8.2.3
terminado==0.18.0
tiktoken==0.5.2
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.3
ufw==0.36.1
ujson==5.9.0
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.1.0
usb-creator==0.3.7
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
virtualenv==20.13.0+ds
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.13
weaviate-client==3.25.3
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
xdg==5
xkit==0.0.0
xyzservices==2023.10.1
yarl==1.9.3
zipp==1.0.0
